Changing directory structure so that 'alignment' and 'genotype' fall under the cpg_workflows/stages directory. Also moving cram_qc.py and gvcf_qc.py into cpg_workflows/stages/alignment and cpg_workflows/stages/genotype respectively to be more in line with these new pipelines 'owning' these stages
