+++ cpg_workflows/stages/alignment/cram_qc.py
@@ -10,7 +10,6 @@ from cpg_utils import Path
 from cpg_utils.config import config_retrieve
 from cpg_utils.hail_batch import get_batch
 from cpg_workflows.filetypes import CramPath
-from cpg_workflows.jobs import somalier
 from cpg_workflows.jobs.multiqc import multiqc
 from cpg_workflows.jobs.picard import picard_collect_metrics, picard_hs_metrics, picard_wgs_metrics
 from cpg_workflows.jobs.samtools import samtools_stats
@@ -53,7 +52,6 @@ def qc_functions() -> list[Qc]:
         return []
 
     qcs = [
-        Qc(func=somalier.extract, outs={'somalier': None}),
         Qc(
             func=verifybamid,
             outs={'verify_bamid': QcOut('.verify-bamid.selfSM', 'verifybamid/selfsm')},
@@ -105,10 +103,7 @@ class CramQC(SequencingGroupStage):
         outs = {}
         for qc in qc_functions():
             for key, out in qc.outs.items():
-                if key == 'somalier':
-                    # Somalier outputs will be written to self.dataset.prefix() / 'cram' / f'{self.id}.cram.somalier' regardless of input cram path.
-                    outs[key] = sequencing_group.make_cram_path().somalier_path
-                elif out:
+                if out:
                     outs[key] = sequencing_group.dataset.prefix() / 'qc' / key / f'{sequencing_group.id}{out.suf}'
         return outs
 
@@ -137,80 +132,7 @@ class CramQC(SequencingGroupStage):
         return self.make_outputs(sequencing_group, data=self.expected_outputs(sequencing_group), jobs=jobs)
 
 
-@stage(required_stages=[CramQC])
-class SomalierPedigree(DatasetStage):
-    """
-    Checks pedigree from CRAM fingerprints.
-    """
-
-    def expected_outputs(self, dataset: Dataset) -> dict[str, Path]:
-        """
-        Return the report for MultiQC, plus putting an HTML into the web bucket.
-        MultiQC expects the following patterns:
-        * *.samples.tsv
-        * *.pairs.tsv
-        https://github.com/ewels/MultiQC/blob/master/multiqc/utils/search_patterns
-        .yaml#L472-L481
-        """
-        if config_retrieve(['workflow', 'skip_qc'], False):
-            return {}
-
-        prefix = dataset.prefix() / 'somalier' / 'cram' / dataset.alignment_inputs_hash()
-        return {
-            'samples': prefix / f'{dataset.name}.samples.tsv',
-            'expected_ped': prefix / f'{dataset.name}.expected.ped',
-            'pairs': prefix / f'{dataset.name}.pairs.tsv',
-            'html': dataset.web_prefix() / 'cram-somalier-pedigree.html',
-            'checks': prefix / f'{dataset.name}-checks.done',
-        }
-
-    def queue_jobs(self, dataset: Dataset, inputs: StageInput) -> StageOutput | None:
-        """
-        Checks calls job from the pedigree module
-        """
-        verifybamid_by_sgid = {}
-        somalier_path_by_sgid = {}
-        for sequencing_group in dataset.get_sequencing_groups():
-            if config_retrieve(['somalier', 'exclude_high_contamination'], False):
-                verify_bamid_path = inputs.as_path(stage=CramQC, target=sequencing_group, key='verify_bamid')
-                if not exists(verify_bamid_path):
-                    logging.warning(
-                        f'VerifyBAMID results {verify_bamid_path} do not exist for '
-                        f'{sequencing_group}, somalier pedigree estimations might be affected',
-                    )
-                else:
-                    verifybamid_by_sgid[sequencing_group.id] = verify_bamid_path
-            somalier_path = inputs.as_path(stage=CramQC, target=sequencing_group, key='somalier')
-            somalier_path_by_sgid[sequencing_group.id] = somalier_path
-
-        html_path = self.expected_outputs(dataset)['html']
-        if base_url := dataset.web_url():
-            html_url = str(html_path).replace(str(dataset.web_prefix()), base_url)
-        else:
-            html_url = None
-
-        if any(sg.pedigree.dad or sg.pedigree.mom for sg in dataset.get_sequencing_groups()):
-            expected_ped_path = dataset.write_ped_file(self.expected_outputs(dataset)['expected_ped'])
-            jobs = somalier.pedigree(
-                b=get_batch(),
-                dataset=dataset,
-                expected_ped_path=expected_ped_path,
-                somalier_path_by_sgid=somalier_path_by_sgid,
-                verifybamid_by_sgid=verifybamid_by_sgid,
-                out_samples_path=self.expected_outputs(dataset)['samples'],
-                out_pairs_path=self.expected_outputs(dataset)['pairs'],
-                out_html_path=html_path,
-                out_html_url=html_url,
-                out_checks_path=self.expected_outputs(dataset)['checks'],
-                job_attrs=self.get_job_attrs(dataset),
-                send_to_slack=True,
-            )
-            return self.make_outputs(dataset, data=self.expected_outputs(dataset), jobs=jobs)
-        else:
-            return self.make_outputs(dataset, skipped=True)
-
-
-@stage(required_stages=[CramQC, SomalierPedigree], analysis_type='qc', analysis_keys=['json'])
+@stage(required_stages=[CramQC], analysis_type='qc', analysis_keys=['json'])
 class CramMultiQC(DatasetStage):
     """
     Run MultiQC to aggregate CRAM QC stats.
@@ -233,8 +155,7 @@ class CramMultiQC(DatasetStage):
 
     def queue_jobs(self, dataset: Dataset, inputs: StageInput) -> StageOutput | None:
         """
-        Call a function from the `jobs` module using inputs from `cramqc`
-        and `somalier` stages.
+        Call a function from the `jobs` module using inputs from `cramqc` stage.
         """
         if config_retrieve(['workflow', 'skip_qc'], False):
             return self.make_outputs(dataset)
@@ -248,16 +169,6 @@ class CramMultiQC(DatasetStage):
             html_url = None
 
         paths = []
-        try:
-            somalier_samples = inputs.as_path(dataset, SomalierPedigree, key='samples')
-            somalier_pairs = inputs.as_path(dataset, SomalierPedigree, key='pairs')
-        except StageInputNotFoundError:
-            pass
-        else:
-            paths = [
-                somalier_samples,
-                somalier_pairs,
-            ]
 
         ending_to_trim = set()  # endings to trim to get sample names
         modules_to_trim_endings = set()
